Fixing some more GL issues
Current working inside DotGraphics and I will need to implement textures soon

+++ Mahjong Engine/Mahjong Engine/Mahjong Engine/FlyingCamera.h
@@ -26,11 +26,10 @@ namespace Engine
 
 		float yaw, pitch;
 		#pragma region Same Thing As
-// float yaw;
-// float pitch;  
-#pragma endregion
-
+		// float yaw;
+		// float pitch;  
+		#pragma endregion
 		bool Editing = false;
-		double lastX, lastY;
+		float lastX, lastY;
 	};
 }
